# Add pool sharing with copy on write