Fixes https://github.com/JuliaData/PooledArrays.jl/issues/54

Could you please have a look at the core of the design proposal. If we are OK with it I will reimplement the rest, add tests, and update documentation.