@bkamins 2021-02-20
Can you have a quick look if now it looks as a right direction? (I will then polish the PR)
I ensures that both `getindex` and `copyto!` will be fast in generic code we have in joining algorithm in DataFrames.jl