@bkamins 2021-02-20
Benchmarks are as follows:

This PR:
```
julia> x = PooledArray(1:10^7);

julia> @time x[1:10^6];
  0.005101 seconds (7 allocations: 7.630 MiB)

julia> @time x[:];
  0.045468 seconds (6 allocations: 76.294 MiB)

julia> y = similar(x);

julia> @time copy!(y, x);
  0.005369 seconds

julia> y = similar(x);

julia> @time copyto!(y, x);
  0.003873 seconds

julia> @time copyto!(y, x);
  0.003631 seconds

julia> y = similar(x);

julia> @time copyto!(y, 1, x, 1, 10^6);
  0.000406 seconds (1 allocation: 16 bytes)

julia> @time copyto!(y, 1, x, 1, 10^6);
  0.000405 seconds (1 allocation: 16 bytes)
```

`main` branch:
```
julia> x = PooledArray(1:10^7);

julia> @time x[1:10^6];
  0.260175 seconds (14 allocations: 288.109 MiB, 3.09% gc time)

julia> @time x[:];
  1.897181 seconds (100 allocations: 580.985 MiB)

julia> y = similar(x);

julia> @time copy!(y, x);
  1.941643 seconds (90 allocations: 542.838 MiB, 3.85% gc time)

julia> y = similar(x);

julia> @time copyto!(y, x);
  1.965071 seconds (90 allocations: 542.838 MiB, 1.09% gc time)

julia> @time copyto!(y, x);
  1.038685 seconds

julia> y = similar(x);

julia> @time copyto!(y, 1, x, 1, 10^6);
  0.095800 seconds (69 allocations: 58.837 MiB)

julia> @time copyto!(y, 1, x, 1, 10^6);
  0.065425 seconds (1 allocation: 16 bytes)
```