@bkamins 2021-02-20
I need to add tests, but other than that this should be good for a review.