@bkamins 2021-02-20
@nalimilan + @quinnj : do you think we should add a separate branch when we detect that Julia is run in a single thread (a quite common case I think) and then we can just ignore everything and never copy `pool` and `invpool`?