@bkamins 2021-02-21
> That makes the code harder to test so it's risky.

OK. Then I will leave it as is for now.