@bkamins 2021-02-21
I will add tests for this functionality later as I have to think how to make them accurate since it is not easy.

What version of PooledArrays.jl this PR should introduce?