@bkamins 2021-02-21
I have added `SubArray` support for most common cases. Hopefully it is OK (I will think more about it).